+++ NEWS.md
@@ -12,6 +12,8 @@
 ### Miscellaneous
 
 - Upon fitting an MMRM, it is checked whether a not reproducible optimization feature of `TMB` is turned on. If so, a warning is issued to the user once per session.
+- `model.matrix` is updated to ensure that the `NA` values are dropped.
+- `model.frame` is updated to ensure that the `na.action` works.
 
 # mmrm 0.3.12